# (shortfin-sd) Adds tooling for performance measurement.