--- a/shortfin/python/shortfin_apps/sd/components/service.py
+++ b/shortfin/python/shortfin_apps/sd/components/service.py
@@ -155,11 +155,25 @@ class GenerateService:
         self.batcher.shutdown()
 
     def __repr__(self):
+        modules = [
+            f"     {key} : {value}" for key, value in self.inference_modules.items()
+        ]
+        params = [
+            f"     {key} : {value}" for key, value in self.inference_parameters.items()
+        ]
         return (
             f"ServiceManager(\n"
-            f"  model_params={self.model_params}\n"
-            f"  inference_modules={self.inference_modules}\n"
-            f"  inference_parameters={self.inference_parameters}\n"
+            f"\n  INFERENCE DEVICES : \n"
+            f"     {self.sysman.ls.devices}\n"
+            f"\n  MODEL PARAMS: \n"
+            f"     {self.model_params}"
+            f"\n  SERVICE PARAMS: \n"
+            f"     fibers per device: {self.fibers_per_device}\n"
+            f"     program isolation mode: {self.prog_isolation}\n"
+            f"\n  INFERENCE MODULES : \n"
+            f"{'\n'.join(modules)}\n"
+            f"\n  INFERENCE PARAMETERS : \n"
+            f"{'\n'.join(params)}\n"
             f")"
         )