… shortfin/python/shortfin_apps/sd/examples/sdxl_request_bs8.json